Grunt type emoji Add documentation for invasion events Update all locale files & include new grunt types Move around some stuff into the new GruntUtils file
