+++ docs/configuration/events/index.rst
@@ -29,6 +29,7 @@ for Events, each with different information:
    egg-events
    raid-events
    weather-events
+   invasion-events
 
 .. _events_dts:
