+++ docs/configuration/events/invasion-events.rst
@@ -29,15 +29,17 @@ Available DTS
 General
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-=============== ==============================
+=============== ==========================================================
 DTS             Description
-=============== ==============================
+=============== ==========================================================
 stop_id         The stop id. Unique per stop.
 stop_name       The name of the stop.
 stop_image      URL to the image of the stop.
-lure_type_id    The id of the active lure.
-lure_type_name  The name of the active lure.
-=============== ==============================
+type_id         The id of the grunt type.
+type_name       The name of the grunt type.
+gender_id       The id of the grunt's gender.
+gender          Gender of the grunt, represented as a single character.
+=============== ==========================================================
 
 
 Location
