+++ locales/de.json
@@ -221,42 +221,56 @@
     }
   },
   "grunt_types": {
-    "04": "Zufall",
-    "05": "Zufall",
-    "06": "K\u00e4fer",
-    "07": "K\u00e4fer",
-    "08": "Geist",
-    "09": "Geist",
-    "10": "Unlicht",
-    "11": "Unlicht",
-    "12": "Drache",
-    "13": "Drache",
-    "14": "Fee",
-    "15": "Fee",
-    "16": "Kampf",
-    "17": "Kampf",
-    "18": "Feuer",
-    "19": "Feuer",
-    "20": "Flug",
-    "21": "Flug",
-    "22": "Pflanze",
-    "23": "Pflanze",
-    "24": "Boden",
-    "25": "Boden",
-    "26": "Eis",
-    "27": "Eis",
-    "28": "Stahl",
-    "29": "Stahl",
-    "30": "Normal",
-    "31": "Normal",
-    "32": "Gift",
-    "33": "Gift",
-    "34": "Psycho",
-    "35": "Psycho",
-    "36": "Gestein",
-    "37": "Gestein",
-    "38": "Wasser",
-    "39": "Wasser"
+    "001": "",
+    "002": "",
+    "003": "",
+    "004": "Zufall",
+    "005": "Zufall",
+    "006": "K\u00e4fer",
+    "007": "K\u00e4fer",
+    "008": "Geist",
+    "009": "Geist",
+    "010": "Unlicht",
+    "011": "Unlicht",
+    "012": "Drache",
+    "013": "Drache",
+    "014": "Fee",
+    "015": "Fee",
+    "016": "Kampf",
+    "017": "Kampf",
+    "018": "Feuer",
+    "019": "Feuer",
+    "020": "Flug",
+    "021": "Flug",
+    "022": "Pflanze",
+    "023": "Pflanze",
+    "024": "Boden",
+    "025": "Boden",
+    "026": "Eis",
+    "027": "Eis",
+    "028": "Stahl",
+    "029": "Stahl",
+    "030": "Normal",
+    "031": "Normal",
+    "032": "Gift",
+    "033": "Gift",
+    "034": "Psycho",
+    "035": "Psycho",
+    "036": "Gestein",
+    "037": "Gestein",
+    "038": "Wasser",
+    "039": "Wasser",
+    "040": "",
+    "041": "Executive Cliff",
+    "042": "Executive Arlo",
+    "043": "Executive Sierra",
+    "044": "Giovanni",
+    "045": "K\u00F6der",
+    "046": "K\u00F6der",
+    "047": "Geist",
+    "048": "Geist",
+    "049": "Elektro",
+    "050": "Elektro"
   },
   "leaders": {
     "000": "Professor Willow",
