+++ locales/en.json
@@ -269,10 +269,10 @@
     "038": 2,
     "039": 1,
     "040": 0,
-    "041": 1,
-    "042": 1,
-    "043": 2,
-    "044": 1,
+    "041": 0,
+    "042": 0,
+    "043": 0,
+    "044": 0,
     "045": 1,
     "046": 2,
     "047": 2,
