+++ locales/es.json
@@ -221,42 +221,56 @@
     }
   },
   "grunt_types": {
-    "04": "Random",
-    "05": "Random",
-    "06": "Bug",
-    "07": "Bug",
-    "08": "Ghost",
-    "09": "Ghost",
-    "10": "Dark",
-    "11": "Dark",
-    "12": "Dragon",
-    "13": "Dragon",
-    "14": "Fairy",
-    "15": "Fairy",
-    "16": "Fighting",
-    "17": "Fighting",
-    "18": "Fire",
-    "19": "Fire",
-    "20": "Flying",
-    "21": "Flying",
-    "22": "Grass",
-    "23": "Grass",
-    "24": "Ground",
-    "25": "Ground",
-    "26": "Ice",
-    "27": "Ice",
-    "28": "Metal",
-    "29": "Metal",
-    "30": "Normal",
-    "31": "Normal",
-    "32": "Poison",
-    "33": "Poison",
-    "34": "Psychic",
-    "35": "Psychic",
-    "36": "Rock",
-    "37": "Rock",
-    "38": "Water",
-    "39": "Water"
+    "001": "",
+    "002": "",
+    "003": "",
+    "004": "Aleatorio",
+    "005": "Aleatorio",
+    "006": "Bicho",
+    "007": "Bicho",
+    "008": "Fantasma",
+    "009": "Fantasma",
+    "010": "Siniestro",
+    "011": "Siniestro",
+    "012": "Drag\u00f3n",
+    "013": "Drag\u00f3n",
+    "014": "Hada",
+    "015": "Hada",
+    "016": "Lucha",
+    "017": "Lucha",
+    "018": "Fuego",
+    "019": "Fuego",
+    "020": "Volador",
+    "021": "Volador",
+    "022": "Planta",
+    "023": "Planta",
+    "024": "Tierra",
+    "025": "Tierra",
+    "026": "Hielo",
+    "027": "Hielo",
+    "028": "Acero",
+    "029": "Acero",
+    "030": "Normal",
+    "031": "Normal",
+    "032": "Veneno",
+    "033": "Veneno",
+    "034": "Ps\u00edquico",
+    "035": "Ps\u00edquico",
+    "036": "Roca",
+    "037": "Roca",
+    "038": "Agua",
+    "039": "Agua",
+    "040": "",
+    "041": "Ejecutivo Cliff",
+    "042": "Ejecutivo Arlo",
+    "043": "Ejecutivo Sierra",
+    "044": "Giovanni",
+    "045": "Se\u00F1uelo",
+    "046": "Se\u00F1uelo",
+    "047": "Fantasma",
+    "048": "Fantasma",
+    "049": "El\u00e9ctrico",
+    "050": "El\u00e9ctrico"
   },
   "leaders": {
     "000": "Profesor Willow",
