+++ locales/fr.json
@@ -225,42 +225,56 @@
     }
   },
   "grunt_types": {
-    "04": "Al\u00e9atoire",
-    "05": "Al\u00e9atoire",
-    "06": "Insecte",
-    "07": "Insecte",
-    "08": "Spectre",
-    "09": "Spectre",
-    "10": "T\u00e9n\u00e8bres",
-    "11": "T\u00e9n\u00e8bres",
-    "12": "Dragon",
-    "13": "Dragon",
-    "14": "F\u00e9e",
-    "15": "F\u00e9e",
-    "16": "Combat",
-    "17": "Combat",
-    "18": "Feu",
-    "19": "Feu",
-    "20": "Vol",
-    "21": "Vol",
-    "22": "Plante",
-    "23": "Plante",
-    "24": "Sol",
-    "25": "Sol",
-    "26": "Glace",
-    "27": "Glace",
-    "28": "Acier",
-    "29": "Acier",
-    "30": "Normal",
-    "31": "Normal",
-    "32": "Poison",
-    "33": "Poison",
-    "34": "Psy",
-    "35": "Psy",
-    "36": "Roche",
-    "37": "Roche",
-    "38": "Eau",
-    "39": "Eau"
+    "001": "",
+    "002": "",
+    "003": "",
+    "004": "Al\u00e9atoire",
+    "005": "Al\u00e9atoire",
+    "006": "Insecte",
+    "007": "Insecte",
+    "008": "Spectre",
+    "009": "Spectre",
+    "010": "T\u00e9n\u00e8bres",
+    "011": "T\u00e9n\u00e8bres",
+    "012": "Dragon",
+    "013": "Dragon",
+    "014": "F\u00e9e",
+    "015": "F\u00e9e",
+    "016": "Combat",
+    "017": "Combat",
+    "018": "Feu",
+    "019": "Feu",
+    "020": "Vol",
+    "021": "Vol",
+    "022": "Plante",
+    "023": "Plante",
+    "024": "Sol",
+    "025": "Sol",
+    "026": "Glace",
+    "027": "Glace",
+    "028": "Acier",
+    "029": "Acier",
+    "030": "Normal",
+    "031": "Normal",
+    "032": "Poison",
+    "033": "Poison",
+    "034": "Psy",
+    "035": "Psy",
+    "036": "Roche",
+    "037": "Roche",
+    "038": "Eau",
+    "039": "Eau",
+    "040": "",
+    "041": "Cliff exécutif",
+    "042": "Arlo exécutif",
+    "043": "Sierra exécutif",
+    "044": "Giovanni",
+    "045": "Leurre",
+    "046": "Leurre",
+    "047": "Spectre",
+    "048": "Spectre",
+    "049": "\u00e9lectrique",
+    "050": "\u00e9lectrique"
   },
   "leaders": {
     "000": "Professeur Saule",
