+++ locales/it.json
@@ -221,42 +221,56 @@
     }
   },
   "grunt_types": {
-    "04": "Random",
-    "05": "Random",
-    "06": "Bug",
-    "07": "Bug",
-    "08": "Ghost",
-    "09": "Ghost",
-    "10": "Dark",
-    "11": "Dark",
-    "12": "Dragon",
-    "13": "Dragon",
-    "14": "Fairy",
-    "15": "Fairy",
-    "16": "Fighting",
-    "17": "Fighting",
-    "18": "Fire",
-    "19": "Fire",
-    "20": "Flying",
-    "21": "Flying",
-    "22": "Grass",
-    "23": "Grass",
-    "24": "Ground",
-    "25": "Ground",
-    "26": "Ice",
-    "27": "Ice",
-    "28": "Metal",
-    "29": "Metal",
-    "30": "Normal",
-    "31": "Normal",
-    "32": "Poison",
-    "33": "Poison",
-    "34": "Psychic",
-    "35": "Psychic",
-    "36": "Rock",
-    "37": "Rock",
-    "38": "Water",
-    "39": "Water"
+    "001": "",
+    "002": "",
+    "003": "",
+    "004": "Casuale",
+    "005": "Casuale",
+    "006": "Roccia",
+    "007": "Roccia",
+    "008": "Fantasma",
+    "009": "Fantasma",
+    "010": "Scuro",
+    "011": "Scuro",
+    "012": "Drago",
+    "013": "Drago",
+    "014": "Fata",
+    "015": "Fata",
+    "016": "Combattimento",
+    "017": "Combattimento",
+    "018": "Fuoco",
+    "019": "Fuoco",
+    "020": "Volare",
+    "021": "Volare",
+    "022": "Erba",
+    "023": "Erba",
+    "024": "Terra",
+    "025": "Terra",
+    "026": "Ghiaccio",
+    "027": "Ghiaccio",
+    "028": "Acciaio",
+    "029": "Acciaio",
+    "030": "Normale",
+    "031": "Normale",
+    "032": "Veleno",
+    "033": "Veleno",
+    "034": "Psichico",
+    "035": "Psichico",
+    "036": "Rock",
+    "037": "Rock",
+    "038": "Acqua",
+    "039": "Acqua",
+    "040": "",
+    "041": "Esecutivo Cliff",
+    "042": "Esecutivo Arlo",
+    "043": "Esecutivo Sierra",
+    "044": "Giovanni",
+    "045": "Esca",
+    "046": "Esca",
+    "047": "Fantasma",
+    "048": "Fantasma",
+    "049": "Elettrico",
+    "050": "Elettrico"
   },
   "leaders": {
     "000": "Professor Willow",
