+++ locales/ko.json
@@ -225,42 +225,56 @@
     }
   },
   "grunt_types": {
-    "04": "Random",
-    "05": "Random",
-    "06": "Bug",
-    "07": "Bug",
-    "08": "Ghost",
-    "09": "Ghost",
-    "10": "Dark",
-    "11": "Dark",
-    "12": "Dragon",
-    "13": "Dragon",
-    "14": "Fairy",
-    "15": "Fairy",
-    "16": "Fighting",
-    "17": "Fighting",
-    "18": "Fire",
-    "19": "Fire",
-    "20": "Flying",
-    "21": "Flying",
-    "22": "Grass",
-    "23": "Grass",
-    "24": "Ground",
-    "25": "Ground",
-    "26": "Ice",
-    "27": "Ice",
-    "28": "Metal",
-    "29": "Metal",
-    "30": "Normal",
-    "31": "Normal",
-    "32": "Poison",
-    "33": "Poison",
-    "34": "Psychic",
-    "35": "Psychic",
-    "36": "Rock",
-    "37": "Rock",
-    "38": "Water",
-    "39": "Water"
+    "001": "",
+    "002": "",
+    "003": "",
+    "004": "\uC784\uC758\uC758",
+    "005": "\uC784\uC758\uC758",
+    "006": "\ubc84\uadf8",
+    "007": "\ubc84\uadf8",
+    "008": "\uc720\ub839",
+    "009": "\uc720\ub839",
+    "010": "\uc5b4\ub460",
+    "011": "\uc5b4\ub460",
+    "012": "\ub4dc\ub798\uace4",
+    "013": "\ub4dc\ub798\uace4",
+    "014": "\uc694\uc815",
+    "015": "\uc694\uc815",
+    "016": "\ud30c\uc774\ud305",
+    "017": "\ud30c\uc774\ud305",
+    "018": "\ubd88",
+    "019": "\ubd88",
+    "020": "\ud50c\ub77c\uc789",
+    "021": "\ud50c\ub77c\uc789",
+    "022": "\uc794\ub514",
+    "023": "\uc794\ub514",
+    "024": "\ubc14\ub2e5",
+    "025": "\ubc14\ub2e5",
+    "026": "\uc5bc\uc74c",
+    "027": "\uc5bc\uc74c",
+    "028": "\uc2a4\ud2f8",
+    "029": "\uc2a4\ud2f8",
+    "030": "\ubcf4\ud1b5",
+    "031": "\ubcf4\ud1b5",
+    "032": "\ub3c5",
+    "033": "\ub3c5",
+    "034": "\uc2ec\ub839",
+    "035": "\uc2ec\ub839",
+    "036": "\ub85d",
+    "037": "\ub85d",
+    "038": "\ubb3c",
+    "039": "\ubb3c",
+    "040": "",
+    "041": "\uD589\uC815\uBD80 Cliff",
+    "042": "\uD589\uC815\uBD80 Arlo",
+    "043": "\uD589\uC815\uBD80 Sierra",
+    "044": "Giovanni",
+    "045": "\uC720\uC778 \uBABB",
+    "046": "\uC720\uC778 \uBABB",
+    "047": "\uc720\ub839",
+    "048": "\uc720\ub839",
+    "049": "\uc804\uae30",
+    "050": "\uc804\uae30"
   },
   "leaders": {
     "000": "\uc70c\ub85c\uc6b0\ubc15\uc0ac",
