+++ locales/pt.json
@@ -225,42 +225,56 @@
     }
   },
   "grunt_types": {
-    "04": "Random",
-    "05": "Random",
-    "06": "Bug",
-    "07": "Bug",
-    "08": "Ghost",
-    "09": "Ghost",
-    "10": "Dark",
-    "11": "Dark",
-    "12": "Dragon",
-    "13": "Dragon",
-    "14": "Fairy",
-    "15": "Fairy",
-    "16": "Fighting",
-    "17": "Fighting",
-    "18": "Fire",
-    "19": "Fire",
-    "20": "Flying",
-    "21": "Flying",
-    "22": "Grass",
-    "23": "Grass",
-    "24": "Ground",
-    "25": "Ground",
-    "26": "Ice",
-    "27": "Ice",
-    "28": "Metal",
-    "29": "Metal",
-    "30": "Normal",
-    "31": "Normal",
-    "32": "Poison",
-    "33": "Poison",
-    "34": "Psychic",
-    "35": "Psychic",
-    "36": "Rock",
-    "37": "Rock",
-    "38": "Water",
-    "39": "Water"
+    "001": "",
+    "002": "",
+    "003": "",
+    "004": "Aleat\u00F3rio",
+    "005": "Aleat\u00F3rio",
+    "006": "Bicho",
+    "007": "Bicho",
+    "008": "Fantasma",
+    "009": "Fantasma",
+    "010": "Escuro",
+    "011": "Escuro",
+    "012": "Drag\u00e3o",
+    "013": "Drag\u00e3o",
+    "014": "Fada",
+    "015": "Fada",
+    "016": "Lutar",
+    "017": "Lutar",
+    "018": "Fogo",
+    "019": "Fogo",
+    "020": "Voando",
+    "021": "Voando",
+    "022": "Grama",
+    "023": "Grama",
+    "024": "Terra",
+    "025": "Terra",
+    "026": "Gelo",
+    "027": "Gelo",
+    "028": "A\u00e7o",
+    "029": "A\u00e7o",
+    "030": "Normal",
+    "031": "Normal",
+    "032": "Veneno",
+    "033": "Veneno",
+    "034": "Ps\u00edquico",
+    "035": "Ps\u00edquico",
+    "036": "Rocha",
+    "037": "Rocha",
+    "038": "\u00c1gua",
+    "039": "\u00c1gua",
+    "040": "",
+    "041": "Executive Cliff",
+    "042": "Executive Arlo",
+    "043": "Executive Sierra",
+    "044": "Giovanni",
+    "045": "Chamariz",
+    "046": "Chamariz",
+    "047": "Fantasma",
+    "048": "Fantasma",
+    "049": "El\u00e9trico",
+    "050": "El\u00e9trico"
   },
   "leaders": {
     "000": "Professor Willow",
