+++ locales/zh_hk.json
@@ -221,42 +221,56 @@
     }
   },
   "grunt_types": {
-    "04": "Random",
-    "05": "Random",
-    "06": "Bug",
-    "07": "Bug",
-    "08": "Ghost",
-    "09": "Ghost",
-    "10": "Dark",
-    "11": "Dark",
-    "12": "Dragon",
-    "13": "Dragon",
-    "14": "Fairy",
-    "15": "Fairy",
-    "16": "Fighting",
-    "17": "Fighting",
-    "18": "Fire",
-    "19": "Fire",
-    "20": "Flying",
-    "21": "Flying",
-    "22": "Grass",
-    "23": "Grass",
-    "24": "Ground",
-    "25": "Ground",
-    "26": "Ice",
-    "27": "Ice",
-    "28": "Metal",
-    "29": "Metal",
-    "30": "Normal",
-    "31": "Normal",
-    "32": "Poison",
-    "33": "Poison",
-    "34": "Psychic",
-    "35": "Psychic",
-    "36": "Rock",
-    "37": "Rock",
-    "38": "Water",
-    "39": "Water"
+    "001": "",
+    "002": "",
+    "003": "",
+    "004": "\u968F\u673A",
+    "005": "\u968F\u673A",
+    "006": "\u87f2",
+    "007": "\u87f2",
+    "008": "\u5e7d\u9748",
+    "009": "\u5e7d\u9748",
+    "010": "\u60e1",
+    "011": "\u60e1",
+    "012": "\u9f8d",
+    "013": "\u9f8d",
+    "014": "\u5996\u7cbe",
+    "015": "\u5996\u7cbe",
+    "016": "\u683c\u9b25",
+    "017": "\u683c\u9b25",
+    "018": "\u706b",
+    "019": "\u706b",
+    "020": "\u98db\u884c",
+    "021": "\u98db\u884c",
+    "022": "\u8349",
+    "023": "\u8349",
+    "024": "\u5730\u9762",
+    "025": "\u5730\u9762",
+    "026": "\u51b0",
+    "027": "\u51b0",
+    "028": "\u92fc",
+    "029": "\u92fc",
+    "030": "\u4e00\u822c",
+    "031": "\u4e00\u822c",
+    "032": "\u6bd2",
+    "033": "\u6bd2",
+    "034": "\u8d85\u80fd\u529b",
+    "035": "\u8d85\u80fd\u529b",
+    "036": "\u5ca9\u77f3",
+    "037": "\u5ca9\u77f3",
+    "038": "\u6c34",
+    "039": "\u6c34",
+    "040": "",
+    "041": "\u884C\u653F\u4EBA\u5458 Cliff",
+    "042": "\u884C\u653F\u4EBA\u5458 Arlo",
+    "043": "\u884C\u653F\u4EBA\u5458 Sierra",
+    "044": "Giovanni",
+    "045": "\u8BF1\u9975",
+    "046": "\u8BF1\u9975",
+    "047": "\u5e7d\u9748",
+    "048": "\u5e7d\u9748",
+    "049": "\u96fb",
+    "050": "\u96fb"
   },
   "leaders": {
     "000": "Professor Willow (\u7dad\u7f85\u535a)",
